Put buttons on dashboard and seperate the styles from main into their own file
