+++ App/Components/Dashboard.js
@@ -1,33 +1,29 @@
 
 var React = require('react-native');
+var styles = require('../Styles/stylessheet');
 
 var {
-  Text,
+  AlertIOS,
+  ActivityIndicatorIos,
+  ListView,
   View,
+  Text,
+  TextInput,
+  TouchableHighlight,
   StyleSheet
 } = React;
-
-var styles = StyleSheet.create({
-  container: {
-    marginTop: 65,
-    flex: 1
-  },
-  image: {
-    height: 350,
-  },
-  buttonText: {
-    fontSize: 24,
-    color: 'white',
-    alignSelf: 'center'
-  }
-});
 
 class Dashboard extends React.Component{
 render(){
   return (
-      <View style={styles.container}>
-        <Text> This is the dashboard </Text>
-        <Text> {JSON.stringify(this.props.userInfo)} </Text>
+      <View style={styles.mainContainer}>
+        <TouchableHighlight style={styles.button}>
+          <Text style={styles.buttonText} >MENU for {this.props.barDetails.name}</Text>
+        </TouchableHighlight>
+        <TouchableHighlight style={styles.button}>
+          <Text style={styles.buttonText}>Dealers Choice</Text>
+        </TouchableHighlight>
+
       </View>
     )
 }
